% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{build_sample}
\alias{build_sample}
\title{Build a sub-sample of a Parquet Dataframe.}
\usage{
build_sample(
  n_samples = 1e+05,
  bucket_input,
  path_input,
  bucket_output,
  path_output
)
}
\arguments{
\item{n_samples}{A number. Size of the sample to keep.}

\item{bucket_input}{A string. Name of the input bucket on MinIO.}

\item{path_input}{A string. Name of the path where to read in the bucket on MinIO.}

\item{bucket_output}{A string. Name of the output bucket on MinIO.}

\item{path_output}{A string. Name of the path where to write in the bucket on MinIO.}
}
\value{
A string. Indicates if the operation was successful.
}
\description{
Build a sub-sample of a Parquet Dataframe.
}
\examples{
\dontrun{
templateProjetMutualise:::build_sample(n_samples=100000,
                                       bucket_input="donnees-insee",
                                       path_input="diffusion/RP/2016/individu_reg/individu_reg.parquet",
                                       bucket_output="projet-formation",
                                       path_output="diffusion/bonnes-pratiques-r/rp_2016_individu_sample.csv"
                                       )
}

}
