% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{decennie_a_partir_annee}
\alias{decennie_a_partir_annee}
\title{Title}
\usage{
decennie_a_partir_annee(annee)
}
\arguments{
\item{annee}{An integer. The year from which we want to compute the decade.}
}
\value{
An integer. The start year of the decade of the input year.
}
\description{
Title
}
\examples{
decennie_a_partir_annee(2022)

}
