% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_data}
\alias{get_data}
\title{Import a Dataframe from a given path.}
\usage{
get_data(path)
}
\arguments{
\item{path}{A string. Path of the input data.}
}
\value{
A Dataframe.
}
\description{
Import a Dataframe from a given path.
}
\examples{
url <- "http://minio.lab.sspcloud.fr/projet-formation/diffusion/bonnes-pratiques-r/rp_2016_individu_sample.csv"
df_sample_rp <- get_data(path=url)

}
