% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stats_agregees}
\alias{stats_agregees}
\title{Title}
\usage{
stats_agregees(x, statistic = "moyenne", ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{statistic}{A string. The statistic to compute.}

\item{...}{Other parameters to pass to the statistic function.}
}
\value{
A number.
}
\description{
Title
}
\examples{
url <- "http://minio.lab.sspcloud.fr/projet-formation/diffusion/bonnes-pratiques-r/rp_2016_individu_sample.csv"
df_sample_rp <- get_data(path = url)
ages_femme <- as.numeric(df_sample_rp$aged[df_sample_rp$sexe == 2])
stats_agregees(x = ages_femme, statistic = "moyenne")

}
